fix: users query repository, delete all-data controller

+++ src/features/blogs/posts/api/posts.controller.ts
@@ -47,6 +47,9 @@ export class PostsController {
 
   @Get('/:id')
   public async getPostsById(@Param('id') id: string) {
+    if (!id) {
+      throw new NotFoundException(`Blog id is required`);
+    }
     const post = await this.postQueryRepository.findPostsAndMap(id);
 
     if (!post) {
@@ -59,12 +62,24 @@ export class PostsController {
   @Put('/:id')
   @HttpCode(204)
   public async putPost(@Body() body: PostInputModel, @Param('id') id: string) {
-    return this.postsService.putPostById(body, id);
+    if (!id) {
+      throw new NotFoundException(`Blog id is required`);
+    }
+    const result = await this.postsService.putPostById(body, id);
+
+    if (!result) {
+      throw new NotFoundException(`post with id ${id} not found`);
+    }
+
+    return;
   }
 
   @Delete('/:id')
   @HttpCode(204)
   public async deleteUser(@Param('id') id: string) {
+    if (!id) {
+      throw new NotFoundException(`Blog id is required`);
+    }
     const result = await this.postsService.deletePost(id);
 
     if (!result) {
@@ -79,6 +94,18 @@ export class PostsController {
     @Param('postId') postId: string,
     @Query() query: { [key: string]: string | undefined },
   ) {
-    return this.postQueryRepository.getPostsComments(query, postId);
+    if (!postId) {
+      throw new NotFoundException(`Blog id is required`);
+    }
+    const result = await this.postQueryRepository.getPostsComments(
+      query,
+      postId,
+    );
+
+    if (!result || result.items.length === 0) {
+      throw new NotFoundException(`Comments with post id ${postId} not found`);
+    }
+
+    return result;
   }
 }
